refactor: move info_plist, mobile_provision to parent directory
